organize dataset layers into separate dataset folder
